network_monitor/requirements.txt
# SIEM Network Monitor - Dependencies
# Python 3.11+

# SNMP
pysnmp==4.4.12
pysnmp-mibs==0.1.6
pysmi==0.3.4

# Async SNMP (better performance)
puresnmp==2.0.0

# Syslog
syslog-rfc5424-parser==1.2.3

# Network utilities
scapy==2.5.0
netaddr==0.9.0

# Database
pyodbc==5.0.1
sqlalchemy==2.0.23

# HTTP client
aiohttp==3.9.1
requests==2.31.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging
structlog==23.2.0
colorlog==6.8.0

# Performance monitoring
psutil==5.9.6

# Scheduling
apscheduler==3.10.4
